requirements.txt
google-generativeai>=0.8.0
earthengine-api>=1.0.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
fastapi>=0.115.0
uvicorn>=0.32.0
websockets>=14.0
python-dotenv>=1.0.0
pydantic>=2.10.0
